requirements.txt
boto >= 2.42.0
boto3 >= 1.4.0
click >= 6.6
configparser >= 3.5.0
Jinja2 >= 2.8
MarkupSafe >= 0.23
python-dateutil >= 2.5.3
